+++ src/services/upload/index.ts
@@ -37,6 +37,16 @@ export interface UploadProgressInfo {
   remainingTime?: number;
 }
 
+// 定义图片上传返回的 data 结构
+export interface UploadImageData {
+  fileUrl: string;
+  fileHash: string;
+  processedFileName: string;
+  originalMimeType: string;
+  processedMimeType: string;
+  imageKitFileId: string;
+}
+
 /**
  * 上传服务类 - 完全基于你已有的 request 服务
  */
@@ -210,6 +220,34 @@ export class UploadService {
     }
 
     return result.data?.data?.status === 'ok';
+  }
+
+  /**
+   * 上传图片或GIF
+   * @param file 要上传的图片文件
+   * @param token 授权token（可选）
+   * @returns 图片URL字符串
+   */
+  async uploadImage(file: File, token?: string): Promise<string> {
+    const formData = new FormData();
+    formData.append('file', file);
+
+    const headers: Record<string, string> = {};
+
+    if (token) {
+      headers['Authorization'] = token;
+    }
+
+    const { data, error } = await request.post<UploadImageData>('/api/v1/upload/avatar', {
+      params: formData,
+      headers,
+    });
+
+    if (error || !data?.data?.fileUrl) {
+      throw new Error(error || '图片上传失败');
+    }
+
+    return data.data.fileUrl;
   }
 }
